Add: Allow starting propagation from fault tree node - (Task #317)
It is now possible to start a propagation connection by clicking on a
fault tree node.
---
Task #317: Allow starting a propagation from a fault tree node
